Move converters to a dedicated subpackage
We have a dedicated subpackage for commands. And one for ops.
Why not one for converters, too?
