+++ src/main/java/net/imagej/itk/convert/SimpleITKImageDatasetConverter.java
@@ -29,9 +29,10 @@
  * #L%
  */
 
-package net.imagej.itk;
+package net.imagej.itk.convert;
 
 import net.imagej.Dataset;
+import net.imagej.itk.SimpleITKService;
 
 import org.itk.simple.Image;
 import org.scijava.Priority;
